Add cluster bomb prefab and update buildable categories
Introduced a new cluster bomb prefab and associated script, updated SampleScene to include a new 'bombs' buildable category, and adjusted build costs for existing buildables. Added performance test resources, updated rendering and post-processing settings, and moved several scripts to a 'badScripts' directory. Also included a recovery scene and made various asset and configuration updates.

+++ remainsOnTheMoon/Assets/scripts/uiScript/popUpUIScript.cs
@@ -21,7 +21,7 @@ public class popUpUIScript : MonoBehaviour// this script is for the tile popups
 
     public float totalIncome = 0; // money income
     public float totalResources = 0;
-    public int totalpopulation = 0;
+    public float totalpopulation = 0;
     public float totalDevelopment = 0;
     // total of all selected tiles, probs should of though of a better name
